# Inline the runState family of functions.