--- a/src/Control/Effect/State/Internal.hs
+++ b/src/Control/Effect/State/Internal.hs
@@ -31,12 +31,14 @@ instance Effect (State s) where
 --   prop> snd (run (runState a get)) == a
 get :: (Member (State s) sig, Carrier sig m) => m s
 get = send (Get pure)
+{-# INLINEABLE get #-}
 
 -- | Project a function out of the current state value.
 --
 --   prop> snd (run (runState a (gets (applyFun f)))) == applyFun f a
 gets :: (Member (State s) sig, Carrier sig m) => (s -> a) -> m a
 gets f = send (Get (pure . f))
+{-# INLINEABLE gets #-}
 
 -- | Replace the state value with a new value.
 --
@@ -45,6 +47,7 @@ gets f = send (Get (pure . f))
 --   prop> snd (run (runState a (put b *> get))) == b
 put :: (Member (State s) sig, Carrier sig m) => s -> m ()
 put s = send (Put s (pure ()))
+{-# INLINEABLE put #-}
 
 -- | Replace the state value with the result of applying a function to the current state value.
 --   This is strict in the new state.
@@ -54,11 +57,13 @@ modify :: (Member (State s) sig, Carrier sig m) => (s -> s) -> m ()
 modify f = do
   a <- get
   put $! f a
+{-# INLINEABLE modify #-}
 
 -- | Replace the state value with the result of applying a function to the current state value.
 --   This is lazy in the new state; injudicious use of this function may lead to space leaks.
 modifyLazy :: (Member (State s) sig, Carrier sig m) => (s -> s) -> m ()
 modifyLazy f = get >>= put . f
+{-# INLINEABLE modifyLazy #-}
 
 -- $setup
 -- >>> :seti -XFlexibleContexts