--- a/src/Control/Effect/State/Strict.hs
+++ b/src/Control/Effect/State/Strict.hs
@@ -26,19 +26,22 @@ import Prelude hiding (fail)
 --
 --   prop> run (runState a (pure b)) == (a, b)
 runState :: s -> StateC s m a -> m (s, a)
-runState = flip runStateC
+runState s x = runStateC x s
+{-# INLINE[3] runState #-}
 
 -- | Run a 'State' effect, yielding the result value and discarding the final state.
 --
 --   prop> run (evalState a (pure b)) == b
 evalState :: forall s m a . Functor m => s -> StateC s m a -> m a
 evalState s = fmap snd . runState s
+{-# INLINE[3] evalState #-}
 
 -- | Run a 'State' effect, yielding the final state and discarding the return value.
 --
 --   prop> run (execState a (pure b)) == a
 execState :: forall s m a . Functor m => s -> StateC s m a -> m s
 execState s = fmap fst . runState s
+{-# INLINE[3] execState #-}
 
 
 newtype StateC s m a = StateC { runStateC :: s -> m (s, a) }