@robrix 2019-05-01
My best guess: inlining `runState` early like that might mean that it discovers the `runStateC` coercion earlier, and thus exposes more opportunities for simplifying the resulting functions.